Fix Paxos acceptor ballot bug, wire prepare previously_accepted, widen RPC frame msg_size to 32-bit

+++ lib/crow-rpc/src/client/client.cpp
@@ -54,7 +54,7 @@ OutFrame *RpcClient::build_frame(uint64_t request_id, Buffer *control, Buffer *d
     frame->control         = control;
     frame->data            = data;
     if (control != nullptr) {
-        frame->header.msg_size = static_cast<uint16_t>(control->len);
+        frame->header.msg_size = control->len;
     }
     if (data != nullptr) {
         frame->header.data_size = data->len;
